Cleanup/Shrinkage.
1) Eliminate more if's in favour of 'cond || stmt' or
   'cond && stmt' constructs.

2) Global _didnet -> DIDNET, make local _nam a local.

3) Display 'DHCP not supported' error every time user
   specifies 'dhcp' but the client is not present.

4) if if_extra is empty, that's ok, so just assign
   $resp to it whether $resp is empty or not.

5) Check for DIDNET inside donetconfig() instead of
   before each call to donetconfig.

6) Eliminate superfluous 'Aborting' comments, and let
   the user draw their own conclusion from getting the
   'Extract more sets' message.

7) Minor code tweaks like not continually assigning
   $resp to a value in a loop, when all we want is
   the final value, and only displaying 'Skipping'
   message in tape installs if we are actually
   skipping files, etc.

+++ distrib/miniroot/install.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-#	$OpenBSD: install.sh,v 1.106 2002/07/13 13:18:05 krw Exp $
+#	$OpenBSD: install.sh,v 1.107 2002/07/13 16:32:13 krw Exp $
 #	$NetBSD: install.sh,v 1.5.2.8 1996/08/27 18:15:05 gwr Exp $
 #
 # Copyright (c) 1997-2002 Todd Miller, Theo de Raadt, Ken Westerback
@@ -90,9 +90,7 @@ if [ -f /etc/fstab ]; then
 else
 	# Install the shadowed disktab file; lets us write to it for temporary
 	# purposes without mounting the miniroot read-write.
-	if [ -f /etc/disktab.shadow ]; then
-		cp /etc/disktab.shadow /tmp/disktab.shadow
-	fi
+	[ -f /etc/disktab.shadow ] && cp /etc/disktab.shadow /tmp/disktab.shadow
 
 	while : ; do
 		if [ -z "$ROOTDISK" ]; then
